requirements.txt
pandas==2.2.3
openpyxl==3.1.5
python-dateutil

#pip install -r requirements.txt